Added Camera shake (horizontal and vertical)
I'll add zoom in a bit, ran out of time before call

Currently, it triggers a vertical shake on snares and horizontal shake
on kicks, both can be changed easily I just did that as an example.
Also we can change the speed, duration, and intensity without much
difficulty

+++ project/Assets/Code/Manager/GameManager.cs
@@ -22,7 +22,7 @@ public static class GameManager {
 	public static List<EnemyObject> enemies;
 	public static EnemySpawnerController spawnController;
 	public static int score = 0;
-	public static float durationForHold = 500f;
+	public static int durationForHold = 500;
 	public static GameTimer gameTimer;
 	
 	static PlayerConfig pConfig = new PlayerConfig();
